# [SPARK-34862][SQL] Support nested column in ORC vectorized reader