--- a/sql/core/src/main/java/org/apache/spark/sql/execution/datasources/orc/OrcColumnVectorUtils.java
+++ b/sql/core/src/main/java/org/apache/spark/sql/execution/datasources/orc/OrcColumnVectorUtils.java
@@ -1,0 +1,54 @@
+package org.apache.spark.sql.execution.datasources.orc;
+
+import org.apache.hadoop.hive.ql.exec.vector.*;
+
+import org.apache.spark.sql.types.*;
+
+/**
+ * Utility class for {@link OrcColumnVector}.
+ */
+class OrcColumnVectorUtils {
+
+  /**
+   * Convert a Hive's {@link ColumnVector} to a Spark's {@link OrcColumnVector}.
+   *
+   * @param type The data type of column vector
+   * @param vector Hive's column vector
+   * @return Spark's column vector
+   */
+  static OrcColumnVector toOrcColumnVector(DataType type, ColumnVector vector) {
+    if (vector instanceof LongColumnVector ||
+      vector instanceof DoubleColumnVector ||
+      vector instanceof BytesColumnVector ||
+      vector instanceof DecimalColumnVector ||
+      vector instanceof TimestampColumnVector) {
+      return new OrcAtomicColumnVector(type, vector);
+    } else if (vector instanceof StructColumnVector) {
+      StructColumnVector structVector = (StructColumnVector) vector;
+      OrcColumnVector[] fields = new OrcColumnVector[structVector.fields.length];
+      int ordinal = 0;
+      for (StructField f : ((StructType) type).fields()) {
+        fields[ordinal] = toOrcColumnVector(f.dataType(), structVector.fields[ordinal]);
+        ordinal++;
+      }
+      return new OrcStructColumnVector(type, vector, fields);
+    } else if (vector instanceof ListColumnVector) {
+      ListColumnVector listVector = (ListColumnVector) vector;
+      OrcColumnVector dataVector = toOrcColumnVector(
+        ((ArrayType) type).elementType(), listVector.child);
+      return new OrcArrayColumnVector(
+        type, vector, dataVector, listVector.offsets, listVector.lengths);
+    } else if (vector instanceof MapColumnVector) {
+      MapColumnVector mapVector = (MapColumnVector) vector;
+      MapType mapType = (MapType) type;
+      OrcColumnVector keysVector = toOrcColumnVector(mapType.keyType(), mapVector.keys);
+      OrcColumnVector valuesVector = toOrcColumnVector(mapType.valueType(), mapVector.values);
+      return new OrcMapColumnVector(
+        type, vector, keysVector, valuesVector, mapVector.offsets, mapVector.lengths);
+    } else {
+      throw new IllegalArgumentException(
+        String.format("OrcColumnVectorUtils.toOrcColumnVector should not take %s as type " +
+          "and %s as vector", type, vector));
+    }
+  }
+}