@c21 2021-03-26
The unit test failed with MiMa tests

> abstract method getBoolean(Int)Boolean in class org.apache.spark.sql.vectorized.ColumnVector does not have a correspondent in current version

However in this PR, the class `org.apache.spark.sql.vectorized.ColumnVector` is not changed at all. I am still checking why this test is failing.