@cloud-fan 2021-03-30
For orc files without nested schema, do we observe perf regression after this PR?